Fixed some build warnings
git-svn-id: svn://mirror.orbiter-radio.co.uk/D3D9client/trunk@277 c0f5cd98-1510-604a-8459-bf7538b8c9c2

+++ Orbitersdk/D3D9Client/VPlanet.cpp
@@ -681,7 +681,7 @@ D3DXVECTOR3 vPlanet::GetSunLightColor(VECTOR3 vPos, float fAmbient, float fGloba
 	double fDPS = max(0.34, dotp(unit(vPos), sundir));	
 	double fDns = exp2(-fAlt * prm.InvSclHeight);
 
-	return exp2(-vOutTotSun * (fDns * AngleCoEff(fDPS)));
+	return exp2(-vOutTotSun * float(fDns * AngleCoEff(fDPS)));
 }
